refactor: extract time range picker into reusable componen
