+++ src/routes/page/conversation/+page.svelte
@@ -19,11 +19,11 @@
 	import TablePagination from '$lib/common/shared/TablePagination.svelte';
 	import LoadingToComplete from '$lib/common/spinners/LoadingToComplete.svelte';
 	import Select from '$lib/common/dropdowns/Select.svelte';
+	import TimeRangePicker from '$lib/common/shared/TimeRangePicker.svelte';
 	import { getAgentOptions } from '$lib/services/agent-service';
 	import { utcToLocal } from '$lib/helpers/datetime';
 	import { ConversationChannel, TimeRange } from '$lib/helpers/enums';
-	import { TIME_RANGE_OPTIONS, CUSTOM_DATE_RANGE } from '$lib/helpers/constants';
-	import { clickoutsideDirective } from '$lib/helpers/directives';
+	import { CUSTOM_DATE_RANGE } from '$lib/helpers/constants';
 	import {
 		getConversations,
 		deleteConversation,
@@ -41,76 +41,11 @@
 	const firstPage = 1;
 	const pageSize = 15;
 
-	// Get today's date in YYYY-MM-DD format
-	const getTodayStr = () => {
-		const d = new Date();
-		return d.getFullYear() + '-' + String(d.getMonth() + 1).padStart(2, '0') + '-' + String(d.getDate()).padStart(2, '0');
-	};
-
-	// Get yesterday's date in YYYY-MM-DD format
-	const getYesterdayStr = () => {
-		const d = new Date();
-		d.setDate(d.getDate() - 1);
-		return d.getFullYear() + '-' + String(d.getMonth() + 1).padStart(2, '0') + '-' + String(d.getDate()).padStart(2, '0');
-	};
-
-	// Format date for display (YYYY-MM-DD -> MM/DD/YYYY)
-	const formatDateForDisplay = (/** @type {string} */ dateStr) => {
-		if (!dateStr) return '';
-		const [year, month, day] = dateStr.split('-');
-		return `${month}/${day}/${year}`;
-	};
-
-	// Get time range display text
-	const getTimeRangeDisplayText = () => {
-		if (searchOption.timeRange === CUSTOM_DATE_RANGE) {
-			if (searchOption.startDate && searchOption.endDate) {
-				const start = formatDateForDisplay(searchOption.startDate);
-				const end = formatDateForDisplay(searchOption.endDate);
-				if (start === end) {
-					return start;
-				}
-				return `${start} - ${end}`;
-			}
-			return 'Custom';
-		}
-		// Find the label for the selected time range
-		const selected = presetTimeRangeOptions.find(x => x.value === searchOption.timeRange);
-		return selected ? selected.label : '';
-	};
-
 	/** @type {boolean} */
 	let isLoading = false;
 	let isComplete = false;
 	let showStateSearch = false;
 	let isPageMounted = false;
-
-	/** @type {boolean} */
-	let showDatePicker = false;
-
-	/** @type {HTMLDivElement | null} */
-	let datePickerRef = null;
-
-	/** @type {string} */
-	let timeRangeDisplayText = '';
-
-	// Update time range display text reactively
-	$: {
-		if (searchOption.timeRange === CUSTOM_DATE_RANGE && searchOption.startDate && searchOption.endDate) {
-			const start = formatDateForDisplay(searchOption.startDate);
-			const end = formatDateForDisplay(searchOption.endDate);
-			if (start === end) {
-				timeRangeDisplayText = start;
-			} else {
-				timeRangeDisplayText = `${start} - ${end}`;
-			}
-		} else if (searchOption.timeRange === CUSTOM_DATE_RANGE) {
-			timeRangeDisplayText = 'Custom';
-		} else {
-			const selected = presetTimeRangeOptions.find(x => x.value === searchOption.timeRange);
-			timeRangeDisplayText = selected ? selected.label : '';
-		}
-	}
 
     /** @type {import('$commonTypes').PagedItems<import('$conversationTypes').ConversationModel>} */
     let conversations = { count: 0, items: [] };
@@ -143,14 +78,6 @@
 		{ value: k.toLowerCase(), label: v }
 	));
 
-	// Preset time range options
-	const presetTimeRangeOptions = TIME_RANGE_OPTIONS.map(x => ({
-			label: x.label,
-			value: x.value
-		}));
-
-	/** @type {string} */
-	let datePickerTab = 'relative';
 
 	/** @type {{ startTime: string | null, endTime: string | null }} */
 	let innerTimeRange = {
@@ -577,154 +504,20 @@
 							on:input={e => changeOption(e, "tags")}
 						/>
 					</Col>
-					<Col lg="2" class="position-relative">
-						<button
-							type="button"
-							class="form-control text-start d-flex align-items-center justify-content-between"
-							on:click={() => {
-								showDatePicker = !showDatePicker;
-								if (showDatePicker) {
-									// If custom date is selected, switch to custom tab; otherwise use relative tab
-									datePickerTab = searchOption.timeRange === CUSTOM_DATE_RANGE ? 'custom' : 'relative';
-								}
+					<Col lg="2">
+						<TimeRangePicker
+							bind:timeRange={searchOption.timeRange}
+							bind:startDate={searchOption.startDate}
+							bind:endDate={searchOption.endDate}
+							on:change={(e) => {
+								searchOption.timeRange = e.detail.timeRange;
+								searchOption.startDate = e.detail.startDate;
+								searchOption.endDate = e.detail.endDate;
+								refreshFilter();
+								initFilterPager();
+								getPagedConversations();
 							}}
-							style="cursor: pointer;"
-						>
-							<span>{timeRangeDisplayText || 'Select time range'}</span>
-							<i class="bx bx-chevron-down"></i>
-						</button>
-						{#if showDatePicker}
-							<div
-								bind:this={datePickerRef}
-								use:clickoutsideDirective
-								on:clickoutside={(/** @type {any} */ e) => {
-									if (e.detail && e.detail.targetNode && datePickerRef) {
-										if (!datePickerRef.contains(e.detail.targetNode)) {
-											showDatePicker = false;
-										}
-									}
-								}}
-								class="position-absolute top-100 start-0 mt-1 bg-white border rounded shadow-lg"
-								style="z-index: 1050; min-width: 320px; max-width: 350px;"
-							>
-								<ul class="nav nav-tabs border-bottom mb-0 px-2 pt-2" role="tablist">
-									<li class="nav-item flex-fill" role="presentation">
-										<button
-											class="nav-link fw-semibold {datePickerTab === 'relative' ? 'active text-primary' : 'text-muted'}"
-											type="button"
-											role="tab"
-											style="padding: 0.5rem 0.75rem; {datePickerTab === 'relative' ? 'border-bottom: 2px solid var(--bs-primary) !important; margin-bottom: -1px;' : ''}"
-											on:click={() => datePickerTab = 'relative'}
-										>
-											Relative
-										</button>
-									</li>
-									<li class="nav-item flex-fill" role="presentation">
-										<button
-											class="nav-link fw-semibold {datePickerTab === 'custom' ? 'active text-primary' : 'text-muted'}"
-											type="button"
-											role="tab"
-											style="padding: 0.5rem 0.75rem; {datePickerTab === 'custom' ? 'border-bottom: 2px solid var(--bs-primary) !important; margin-bottom: -1px;' : ''}"
-											on:click={() => {
-												datePickerTab = 'custom';
-												// Set default dates to yesterday and today if not already set
-												if (!searchOption.startDate && !searchOption.endDate) {
-													searchOption.startDate = getYesterdayStr();
-													searchOption.endDate = getTodayStr();
-												}
-											}}
-										>
-											Custom
-										</button>
-									</li>
-								</ul>
-								
-								<div class="p-4">
-									{#if datePickerTab === 'relative'}
-										<div class="d-flex flex-column gap-2" style="max-height: 300px; overflow-y: auto;">
-											{#each presetTimeRangeOptions as option}
-												<button
-													type="button"
-													class="btn btn-sm btn-outline-secondary text-start {searchOption.timeRange === option.value ? 'active' : ''}"
-													on:click={(e) => {
-														e.preventDefault();
-														e.stopPropagation();
-														searchOption.timeRange = option.value;
-														searchOption.startDate = '';
-														searchOption.endDate = '';
-														showDatePicker = false;
-														refreshFilter();
-														initFilterPager();
-														getPagedConversations();
-													}}
-												>
-													{option.label}
-												</button>
-											{/each}
-										</div>
-									{:else if datePickerTab === 'custom'}
-										<div class="mb-3">
-											<label for="start-date-picker" class="form-label small mb-2">From:</label>
-											<Input
-												id="start-date-picker"
-												type="date"
-												bind:value={searchOption.startDate}
-												class="form-control form-control-sm"
-											/>
-										</div>
-										<div class="mb-4">
-											<label for="end-date-picker" class="form-label small mb-2">To:</label>
-											<Input
-												id="end-date-picker"
-												type="date"
-												bind:value={searchOption.endDate}
-												class="form-control form-control-sm"
-											/>
-										</div>
-										<div class="d-flex justify-content-end gap-2 mt-3">
-											<Button
-												color="secondary"
-												size="sm"
-												type="button"
-												on:click={(e) => {
-													e.preventDefault();
-													e.stopPropagation();
-													showDatePicker = false;
-												}}
-											>
-												Cancel
-											</Button>
-											<Button
-												color="primary"
-												size="sm"
-												type="button"
-												on:click={(e) => {
-													e.preventDefault();
-													e.stopPropagation();
-													if (searchOption.startDate) {
-														// If endDate is not provided, default to startDate
-														if (!searchOption.endDate) {
-															searchOption.endDate = searchOption.startDate;
-														}
-														// Force reactivity by reassigning the object
-														searchOption = {
-															...searchOption,
-															timeRange: CUSTOM_DATE_RANGE
-														};
-													}
-													showDatePicker = false;
-													refreshFilter();
-													initFilterPager();
-													getPagedConversations();
-												}}
-											>
-												Confirm
-											</Button>
-										</div>
-									{/if}
-								</div>
-							</div>
-						{/if}
+						/>
 					</Col>
 					<Col lg="1">
 						<Button
